Move CoerceUnsized to module.
